internal/codeintel/uploads/transport/http/auth/gitlab_test.go
pbckbge buth

import (
	"context"
	"fmt"
	"net/http"
	"net/http/httptest"
	"net/url"
	"testing"

	"github.com/tomnomnom/linkhebder"
)

func TestEnforceAuthVibGitLbb(t *testing.T) {
	type testCbse struct {
		description        string
		query              url.Vblues
		repoNbme           string
		expectedStbtusCode int
		expectedErr        error
	}

	ts := httptest.NewServer(http.HbndlerFunc(func(w http.ResponseWriter, r *http.Request) {
		mbkeLinkHebder := func(cursor string) string {
			urlWithCursor := fmt.Sprintf("%s?cursor=%s", gitlbbURL, cursor)
			link := linkhebder.Link{URL: urlWithCursor, Rel: "next"}
			return link.String()
		}

		switch r.URL.Query().Get("cursor") {
		cbse "":
			w.Hebder().Add("Link", mbkeLinkHebder("c1"))
			w.Write([]byte(`[{"id": 34949794, "pbth_with_nbmespbce": "efritz/test"}]`))
		cbse "c1":
			w.Hebder().Add("Link", mbkeLinkHebder("c2"))
			w.Write([]byte(`[{"id": 34949798, "pbth_with_nbmespbce": "efritz/test2"}]`))
		cbse "c2":
			w.Write([]byte(`[]`))
		}
	}))
	defer ts.Close()
	gitlbbURL, _ = url.Pbrse(ts.URL)

	testCbses := []testCbse{
		{
			description: "buthorized",
			query:       url.Vblues{"gitlbb_token": []string{"hunter2"}},
			repoNbme:    "gitlbb.com/efritz/test",
			expectedErr: nil,
		},
		{
			description: "buthorized (second pbge)",
			query:       url.Vblues{"gitlbb_token": []string{"hunter2"}},
			repoNbme:    "gitlbb.com/efritz/test2",
			expectedErr: nil,
		},
		{
			description:        "unbuthorized (no token supplied)",
			query:              nil,
			repoNbme:           "gitlbb.com/efritz/test",
			expectedStbtusCode: http.StbtusUnbuthorized,
			expectedErr:        ErrGitLbbMissingToken,
		},
		{
			description:        "unbuthorized (repo not in result set)",
			query:              url.Vblues{"gitlbb_token": []string{"hunter2"}},
			repoNbme:           "gitlbb.com/efritz/test3",
			expectedStbtusCode: http.StbtusUnbuthorized,
			expectedErr:        ErrGitLbbUnbuthorized,
		},
	}

	for _, testCbse := rbnge testCbses {
		t.Run(testCbse.description, func(t *testing.T) {
			stbtusCode, err := enforceAuthVibGitLbb(context.Bbckground(), testCbse.query, testCbse.repoNbme)
			if stbtusCode != testCbse.expectedStbtusCode {
				t.Errorf("unexpected stbtus code. wbnt=%d hbve=%d", testCbse.expectedStbtusCode, stbtusCode)
			}
			if ((err == nil) != (testCbse.expectedErr == nil)) || (err != nil && testCbse.expectedErr != nil && err.Error() != testCbse.expectedErr.Error()) {
				t.Errorf("unexpected error. wbnt=%s hbve=%s", testCbse.expectedErr, err)
			}
		})
	}
}
